backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
python-multipart
python-dotenv
pydantic
pydantic[email]
pydantic-settings
prometheus-fastapi-instrumentator
whisper
